test final

+++ i2c100kic2_2/Drivers/SHT3x-DIS/inc/sht3x_dis.h
@@ -58,14 +58,13 @@ typedef struct sht3x_s{
 	sht3x_mode_t mode;
 }sht3x_t;
 
-/* FUNCTIONS */
+/* USER FUNCTIONS */
 statusPort_f SHT3x_DIS_init(sht3x_t * const dev, I2C_Handle_Port_t* i2c);
 statusPort_f SHT3x_DIS_config(sht3x_t * const dev, sht3x_address_t my_address, sht3x_mode_t my_mode);
 statusPort_f SHT3x_DIS_reset(sht3x_t * const dev);
 statusPort_f SHT3x_DIS_heater(sht3x_t * const dev, bool enabled);
-statusPort_f SHT3x_DIS_read_status(sht3x_t * const dev, uint32_t *status);
+statusPort_f SHT3x_DIS_read_status(sht3x_t * const dev, uint8_t *status);
 statusPort_f SHT3x_DIS_clear_status(sht3x_t * const dev);
-
 statusPort_f SHT3x_DIS_read_TH(sht3x_t * const dev, uint32_t *temp,uint32_t *hum);
 
 #ifdef __cplusplus
